Splitting Caller class
