+++ BHoM_UI/BHoM_UI.csproj
@@ -127,6 +127,14 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Caller\Input_CRUD.cs" />
+    <Compile Include="Caller\Menu.cs" />
+    <Compile Include="Caller\Read.cs" />
+    <Compile Include="Caller\Compile.cs" />
+    <Compile Include="Caller\RecordError.cs" />
+    <Compile Include="Caller\Write.cs" />
+    <Compile Include="Caller\SetItem.cs" />
+    <Compile Include="Caller\Run.cs" />
     <Compile Include="Components\Engine\External.cs" />
     <Compile Include="Components\Adapter\CreateRequest.cs" />
     <Compile Include="Components\Adapter\CreateAdapter.cs" />
@@ -156,13 +164,12 @@
     <Compile Include="Global\SearchMenu.cs" />
     <Compile Include="Global\SearchMenu_WinForm.cs" />
     <Compile Include="Global\SearchMenu_Wpf.cs" />
-    <Compile Include="Templates\DataAccessor.cs" />
     <Compile Include="Templates\ParamSelectorMenu.cs" />
     <Compile Include="Templates\IItemSelectorMenu.cs" />
     <Compile Include="Templates\IItemSelector.cs" />
     <Compile Include="Templates\MultiChoiceCaller.cs" />
     <Compile Include="Templates\ItemSelector.cs" />
-    <Compile Include="Templates\Caller.cs" />
+    <Compile Include="Caller\Core.cs" />
     <Compile Include="Templates\MethodCaller.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs" />
