+++ BHoM_UI/Caller/RecordError.cs
@@ -1,0 +1,82 @@
+/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using BH.Engine.Reflection;
+using BH.oM.Reflection;
+using BH.oM.UI;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Reflection;
+using BH.Engine.Serialiser;
+using System.Windows.Forms;
+using BH.oM.Base;
+using System.Collections;
+
+namespace BH.UI.Templates
+{
+    public abstract partial class Caller
+    {
+        /*************************************/
+        /**** Public Methods              ****/
+        /*************************************/
+
+        protected static void RecordError(Exception e, string message = "")
+        {
+            if (e.InnerException != null)
+                message += e.InnerException.Message;
+            else
+                message += e.Message;
+            BH.Engine.Reflection.Compute.RecordError(message);
+        }
+
+        /*******************************************/
+
+        protected virtual void RecordExecutionError(Exception e)
+        {
+            string message = "This component failed to run properly.\n- Error: ";
+
+            if (e.InnerException != null)
+                message += e.InnerException.Message;
+            else
+                message += e.Message;
+
+            List<string> stack = e.StackTrace.Split(new char[] { '\n' })
+                .Where(x => x.Contains(" BH."))
+                .Take(2)
+                .ToList();
+
+            if (stack.Count > 0)
+                message += "\n- Occured in " + stack[0].Trim().Substring(2);
+            if (stack.Count > 1)
+                message += "\n     called from" + stack[1].Trim().Substring(2);
+
+            message += "\n- Are you sure you have the correct type of inputs? Check their description for more details.";
+
+            Engine.Reflection.Compute.RecordError(message);
+        }
+
+        /*************************************/
+    }
+}
+
