+++ UI_Engine/Query/AreMatching.cs
@@ -1,0 +1,59 @@
+/*
+ * This file is part of the Buildings and Habitats object Model (BHoM)
+ * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
+ *
+ * Each contributor holds copyright over their respective contributions.
+ * The project versioning (Git) records all such contribution source information.
+ *                                           
+ *                                                                              
+ * The BHoM is free software: you can redistribute it and/or modify         
+ * it under the terms of the GNU Lesser General Public License as published by  
+ * the Free Software Foundation, either version 3.0 of the License, or          
+ * (at your option) any later version.                                          
+ *                                                                              
+ * The BHoM is distributed in the hope that it will be useful,              
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of               
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the                 
+ * GNU Lesser General Public License for more details.                          
+ *                                                                            
+ * You should have received a copy of the GNU Lesser General Public License     
+ * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
+ */
+
+using BH.Engine.Reflection;
+using BH.oM.UI;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BH.Engine.UI
+{
+    public static partial class Query
+    {
+        /*************************************/
+        /**** Public Methods              ****/
+        /*************************************/
+
+        public static bool AreMatching(List<ParamInfo> newList, List<ParamInfo> oldList)
+        {
+            if (newList.Count != oldList.Count)
+                return false;
+            else
+                return newList.Zip(oldList, (a, b) => a.Name == b.Name).All(x => x);
+
+        }
+
+        /*************************************/
+
+        public static bool AreMatching(List<PropertyInfo> props, List<ParamInfo> oldList)
+        {
+            return oldList.All(x => props.Exists(p => p.Name == x.Name && p.PropertyType == x.DataType));
+        }
+
+        /*************************************/
+    }
+}
+
