+++ UI_Engine/UI_Engine.csproj
@@ -85,8 +85,11 @@
     <Compile Include="Convert\FromProperty.cs" />
     <Compile Include="Compute\ConstructorText.cs" />
     <Compile Include="Compute\Constructor.cs" />
+    <Compile Include="Create\OutputAccessor.cs" />
+    <Compile Include="Create\InputAccessor.cs" />
     <Compile Include="Create\ParamInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Query\AreMatching.cs" />
     <Compile Include="Query\Settings.cs" />
     <Compile Include="Query\Hits.cs" />
     <Compile Include="Query\Items.cs" />
